// apps/web/src/components/country-map-cn.generated.ts
export const CHINA_MAP_BOUNDS = {
  minLongitude: 73.5024,
  maxLongitude: 135.0957,
  minLatitude: 18.1436,
  maxLatitude: 53.5583,
} as const;